add eventbus protocol, prepare multiview architecture
