+++ app/build.gradle
@@ -50,6 +50,7 @@ dependencies {
 
     compile 'com.jakewharton:butterknife:7.0.1'
 
+    compile 'de.greenrobot:eventbus:3.0.0-beta1'
     compile 'de.greenrobot:greendao:1.3.7'
     //uncomment when available on maven central
     //compile 'eu.codlab:mutex_shared:0.9.2'
