+++ app/src/androidTest/java/com/wikaba/ogapp/loaders/FleetEventLoaderTest.java
@@ -22,7 +22,7 @@ package com.wikaba.ogapp.loaders;
 import android.support.test.runner.AndroidJUnit4;
 
 import com.wikaba.ogapp.AgentService;
-import com.wikaba.ogapp.HomeActivity;
+import com.wikaba.ogapp.ui.main.HomeActivity;
 import com.wikaba.ogapp.agent.FleetAndResources;
 import com.wikaba.ogapp.agent.FleetEvent;
 import com.wikaba.ogapp.agent.IntegerMissionMap;
