+++ app/src/main/AndroidManifest.xml
@@ -29,11 +29,11 @@
     <application
         android:name=".ApplicationController"
         android:allowBackup="true"
-        android:icon="@drawable/ic_launcher"
+        android:icon="@drawable/ic_launcher_app"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.main.HomeActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
