+++ app/src/main/java/com/wikaba/ogapp/ui/overview/EventViewHolder.java
@@ -1,0 +1,18 @@
+package com.wikaba.ogapp.ui.overview;
+
+import android.widget.LinearLayout;
+import android.widget.TextView;
+
+/**
+ * Created by kevinleperf on 19/07/15.
+ */
+class EventViewHolder {
+    TextView eta;
+    TextView originCoords;
+    TextView outOrIn;
+    TextView destCoords;
+    TextView missionType;
+    LinearLayout civilShips;
+    LinearLayout combatShips;
+    LinearLayout resources;
+}
